feat: handle processing of large files
